requirements.txt
hydra-core~=1.2.0
lmdb~=1.3.0
matplotlib~=3.5.2
numpy~=1.21.6
pytorch-lightning~=1.6.4
torchcrepe~=0.0.16
tqdm~=4.64.0
git+https://github.com/QosmoInc/neutone_sdk.git#egg=neutone_sdk

--extra-index-url https://download.pytorch.org/whl/cu113
torch==1.11.0+cu113
torchaudio==0.11.0